use Integer pkey

+++ src/main/java/org/dbos/apiary/executor/ApiaryConnection.java
@@ -1,5 +1,5 @@
 package org.dbos.apiary.executor;
 
 public interface ApiaryConnection {
-    FunctionOutput callFunction(String name, long pkey, Object... inputs) throws Exception;
+    FunctionOutput callFunction(String name, int pkey, Object... inputs) throws Exception;
 }
